+++ lib/src/models/models.dart
@@ -21,3 +21,8 @@ export 'usage/usage_response.dart';
 
 // Tile
 export 'tile/tile_url_response.dart';
+
+// Smart Search
+export 'smart_search/opening_hours_entry.dart';
+export 'smart_search/smart_search_result.dart';
+export 'smart_search/smart_search_response.dart';